correct formatting

+++ client/src/components/Nav/Navbar/index.js
@@ -14,26 +14,41 @@ export default class Home extends Component {
     const jwt_decoded = await access_token ? jwtdecode(access_token) : ""; 
     const expired = `${new Date().getTime() > jwt_decoded.exp ? true : false}`;
 
-    switch (expired) { 
-      case 'false': 
-        window.localStorage.clear();
-        this.setState({ loggedIn: false, nav: <><li className="nav">Login</li><li className="nav">Register</li></> }); 
-        break;
-      // If jwt is not expired, validate the user
-      default: 
-        axios.post("/auth/validate", { 
-          access_token: access_token, 
-          uuid: jwt_decoded.uuid, 
-        }).then(res => {
-          switch (res.data.access) {
-            case "approve": 
-              window.localStorage.setItem("access_token", res.data.access_token);
-              this.setState({ loggedIn: true, nav: <li className="nav">Profile</li> }); 
-              break;
-            default: 
-              this.setState({ loggedIn: false, nav: <><li className="nav">Login</li><li className="nav">Register</li></> }); 
-          };
-        }).catch(err => console.log(err));
+    if(expired === 'false') { 
+      window.localStorage.clear();
+      this.setState({ 
+        loggedIn: false, 
+        nav: (
+          <>
+            <li className="nav">Login</li>
+            <li className="nav">Register</li>
+          </> 
+        )
+      }); 
+    } else {
+      axios.post("/auth/validate", { 
+        access_token: access_token, 
+        uuid: jwt_decoded.uuid, 
+      }).then(res => {
+        if(res.data.access === "approve") {
+          window.localStorage.setItem("access_token", res.data.access_token);
+          this.setState({ 
+            loggedIn: true, 
+            nav: (
+              <li className="nav">Profile</li>
+            )
+          }); 
+        } else {
+          this.setState({ 
+            loggedIn: false, nav: (
+              <>
+                <li className="nav">Login</li>
+                <li className="nav">Register</li>
+              </> 
+            )
+          }); 
+        }
+      }).catch(err => console.log(err));
     }
   };
